json-to-docx/requirements.txt
argparse
PyYAML
lxml
httplib2
yamllint

pandas
numpy

oauth2client
PyDrive
pygsheets

progress
tqdm
termcolor
colorama
